+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # sql to java change log
 
+## 2023-12-02
+* fix windows os .Author null
+* support organizing directories by package path 
+* support some built-in functions
+
 ## 2023-11-25
 * support generate java file from templates